requirements.txt
asgiref==3.7.2
attrs==21.4.0
Babel==2.15.0
beautifulsoup4==4.12.2
black==24.3.0
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==4.2.22
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==21.1
django-heroku==0.3.1
django-import-export==3.3.8
django-json-widget==2.0.1
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==6.3.0
django-sesame==3.1
django-unfold==0.22.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.51.0
geographiclib==1.52
google-api-core==2.24.2
google-api-python-client==2.147.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.2
google-cloud-core==2.4.1
google-cloud-translate==3.18.0
googleapis-common-protos==1.56.4
grpcio==1.58.0
grpcio-status==1.48.2
gunicorn==23.0.0
httplib2==0.20.4
hubspot-api-client==9.0.0
idna==3.7
inflection==0.5.1
iniconfig==1.1.1
kiwisolver==1.4.4
MarkupPy==1.14
mypy-extensions==1.0.0
numpy==1.23.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.4
packaging==23.1
pathspec==0.11.1
phonenumbers==8.12.53
pillow==10.3.0
platformdirs==3.5.1
pluggy==1.0.0
proto-plus==1.22.3
protobuf==4.25.8
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==3.0.1
pydot-ng==2.0.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.9.7
sentry-sdk==2.8.0
sib-api-v3-sdk==7.6.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
tablib==3.5.0
tomli==2.0.1
tqdm==4.66.3
twilio==7.16.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.5.0
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
